requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1

